# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SPV API
  version: 1.0.0
  description: API para votos, puntos y analisis de planificacion.
servers:
  - url: http://localhost:8787
paths:
...
  /api/votes:
    post:
      summary: Registrar voto y acreditar puntos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [activityId]
              properties:
                activityId:
                  type: string
                activityScope:
                  type: string
                  default: local
                requestId:
                  type: string
                eventId:
                  type: string
                metadata:
                  type: object
      responses:
        '201':
          description: Voto procesado
        '429':
          description: Limite de votos o rate limit excedido
  /api/votes/count:
    get:
      summary: Contar votos del dia
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: activityId
          schema:
            type: string
      responses:
        '200':
          description: Conteo de votos
  /api/votes/limits:
    get:
      summary: Consultar limites diarios de votos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Limites de usuario
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT